apiforms/requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
distlib==0.3.5
Django==4.1.3
django-crispy-forms==1.14.0
django-jsonform==2.14.0
django-jsonforms==1.1.2
filelock==3.7.1
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud-core==2.3.2
google-cloud-speech==2.15.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
httplib2==0.20.4
idna==3.3
importlib-resources==5.10.0
jsonschema==4.17.0
numpy==1.21.4
oauthlib==3.2.1
opencv-python==4.5.4.60
pipenv==2022.7.24
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
proto-plus==1.22.1
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pyparsing==3.0.9
pyrsistent==0.19.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.1
virtualenv-clone==0.5.7
zipp==3.10.0
